nidas::core::SampleThread nidas::core::SampleBuffer nidas::core::SampleSorter     Changed return of getNumDiscardedSamples to size_t from unsigned int. nidas::dynld::UDPSampleOutput nidas::dynld::SampleOutputStream     Warning message for getNumDiscardedMessages was using %lld format descriptor. Change to %z.

+++ src/nidas/core/SampleSorter.h
@@ -204,7 +204,7 @@ public:
      * Number of samples discarded because of _heapSize > _heapMax
      * and heapBlock == true.
      */
-    unsigned int getNumDiscardedSamples() const
+    size_t getNumDiscardedSamples() const
     {
         return _discardedSamples;
     }
@@ -213,7 +213,7 @@ public:
      * Number of samples discarded because their timetags 
      * were in the future.
      */
-    unsigned int getNumFutureSamples() const
+    size_t getNumFutureSamples() const
     {
         return _realTimeFutureSamples;
     }
